requirements/scripts.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    make upgrade
#
atlassian-python-api==3.8.0
    # via -r requirements/scripts.in
certifi==2020.12.5
    # via requests
chardet==4.0.0
    # via requests
click==7.1.2
    # via
    #   -r requirements/base.txt
    #   code-annotations
code-annotations==1.1.0
    # via
    #   -r requirements/base.txt
    #   -r requirements/scripts.in
deprecated==1.2.12
    # via atlassian-python-api
django-crum==0.7.9
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
django-waffle==2.1.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
django==2.2.19
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/base.txt
    #   code-annotations
    #   django-crum
    #   edx-django-utils
edx-django-utils==3.15.0
    # via -r requirements/base.txt
idna==2.10
    # via requests
jinja2==2.11.3
    # via
    #   -r requirements/base.txt
    #   code-annotations
markupsafe==1.1.1
    # via
    #   -r requirements/base.txt
    #   jinja2
newrelic==6.2.0.156
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
oauthlib==3.1.0
    # via
    #   atlassian-python-api
    #   requests-oauthlib
pbr==5.5.1
    # via
    #   -r requirements/base.txt
    #   stevedore
psutil==5.8.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
python-slugify==4.0.1
    # via
    #   -r requirements/base.txt
    #   code-annotations
pytz==2021.1
    # via
    #   -r requirements/base.txt
    #   django
pyyaml==5.4.1
    # via
    #   -r requirements/base.txt
    #   code-annotations
requests-oauthlib==1.3.0
    # via atlassian-python-api
requests==2.25.1
    # via
    #   atlassian-python-api
    #   requests-oauthlib
six==1.15.0
    # via
    #   -r requirements/base.txt
    #   atlassian-python-api
    #   stevedore
sqlparse==0.4.1
    # via
    #   -r requirements/base.txt
    #   django
stevedore==1.32.0
    # via
    #   -c requirements/constraints.txt
    #   -r requirements/base.txt
    #   code-annotations
    #   edx-django-utils
text-unidecode==1.3
    # via
    #   -r requirements/base.txt
    #   python-slugify
urllib3==1.26.4
    # via requests
wrapt==1.11.0
    # via
    #   -c requirements/constraints.txt
    #   deprecated
